Bump PyO3 version and support Stable Rust
This commit updates retworkx to use the latest version of PyO3 which
enables support for compiling with stable rust (rustc >= 1.39). While
updating the PyO3 version supported this takes the opportunity to update
all the documentation and CI builds to rely on stable rust instead of
nightly. The nightly step was always an extra manual step that detracted
from the contribution process and prevented use in environments where
rustup wasn't allowed to be used.