Move documentation inline and rework documentation (#90)
* Move documentation inline and rework documentation

This commit is the beginning of an overhaul of the documentation for
retworkx. Previously the entire structure of the retworkx documentation
was hand built as rst files. When the project first started this was
fine because there was 1 class, PyDAG and only a few functions that used
PyDAG, but now that the project has started to grow dumping everything
into a single rst file hasn't scaled well. This resulted in several
problems primarily the formatting was all wrong, but also that the
structure was poorly organized. The first step this takes is moving the
rst files to inline docstrings, PyO3 is capable of exporting the rust
docstrings and function signatures to python so we should leverage that
capability, that way interactive help tools (like with ipython and
jupyter) can see the docs for everything, but also we can leverage
sphinx autodoc to help with the structure and formating. From there the
organization of the docs is reworked to split up the PyGraph, PyDiGraph,
and PyDAG into one section and then to put the algorithm functions into
another section.

* Fix sphinx build and update tox.ini

This commit fixes the source files in the docs so that they're built
with the expected structure. Then to facilitate local testing the
tox.ini is updated so that tox -edocs actually works as expected.