Attempt to fix readthedocs build
Since #90 the docs are now autogenerated from the installed retworkx
package. However, the readthedocs builds are not working because they
do not have rustc installed in their build env so it is unable to build
retworkx. This commit attempts to configure readthedocs to install rust
via conda and then use that to build retworkx.