Actually export exception types
There were several custom extension classes being defined in retworkx
that were never being exported to the retworkx module. This left users
in a weird position where a custom exception would be raised from a
retworkx call but they would be unable to import that exception class
from retworkx to catch it. This commit fixes this issue by adding the
exceptions to the module so that python code can use the exceptions. The
tests and docs are updated accordingly to use the newly exported
exception classes instead of the base Exception class (which is the
parent for all of retworkx's exceptions).