Add remove_nodes_from method
This commit adds a new method to PyGraph and PyDiGraph (which includes
PyDAG), remove_nodes_from, which enables removing multiple nodes at
once.

As part of this addition there was an obvious gap in the documentation
and testing for remove_node() for when it was called with an invalid
node index. While the behavior was defined it was never called out or
tested for. This commit took the opportunity to do this since it was
tangentially related to the new methods being added here.