Fix copy and paste issues with tests
The previous commit copy and pasted the PyDAG/PyDiGraph tests for the new
method into the PyGraph tests but neglected to update them to actually
use a PyGraph instead. This results in a coverage gap as we ended up
running the PyDAG/PyDiGraph version twice and never running it for
PyGraph. This commit fixes the oversight and updates the graph test
methods to actually use a PyGraph.