Add remove_nodes_from method (#95)
* Add remove_nodes_from method

This commit adds a new method to PyGraph and PyDiGraph (which includes
PyDAG), remove_nodes_from, which enables removing multiple nodes at
once.

As part of this addition there was an obvious gap in the documentation
and testing for remove_node() for when it was called with an invalid
node index. While the behavior was defined it was never called out or
tested for. This commit took the opportunity to do this since it was
tangentially related to the new methods being added here.

* Fix copy and paste issues with tests

The previous commit copy and pasted the PyDAG/PyDiGraph tests for the new
method into the PyGraph tests but neglected to update them to actually
use a PyGraph instead. This results in a coverage gap as we ended up
running the PyDAG/PyDiGraph version twice and never running it for
PyGraph. This commit fixes the oversight and updates the graph test
methods to actually use a PyGraph.