Add dijkstra shortest_path_lengths functions
This commit adds a new function for finding the shortest path lengths
from a node. This is built using the dijkstra's algorithm function from
the upstream petgraph lib, however because the data for the graphs being
traversed are PyObjects we needed to be able to handle python exceptions
gracefully (instead of panicking) which was not possible with the upstream
dijkstra function. So this function has been forked into retworkx and
modified to hand PyResult<> wrapped returns from the callables.