Avoid use of python for in/out_edges
This commit reworks the generation of the output list from the
out_edges() and in_edges() methods for PyDiGraph/PyDAG. Previously these
methods were manually looping over the edges and constructing a Vec of
PyObjects from a tuple. This ends up being a performance bottleneck
because we need to call out to python on each iteration to convert the
tuple to a python object. Then we call out to python again at the end of
the method to convert the Vec to to a PyList and return that. Instead
this changes the logic to just generate a Vec<(source, target, weight)>
using map() over the iterator and return that. This relies on PyO3 to do
the conversion for us which ends up being faster because it happens once
at the end instead of on every iteration of the loop.