Update coverage rustflags
This commit updates the flags used for generating coverage data with
grcov based on the latest README for grcov. [1] The last time these were
updated in #72 was to remove a flag that was being removed from rust and
causing the job to fail. However, that commit failed to add equivalent
flags which would perform the same functionality. This commit fixes that
oversight so we should have more reliable coverage collection.

[1] https://github.com/mozilla/grcov/blob/master/README.md