Remove edges close to probability tests
The 2 edge_close_to_probability test methods were copied from networkx's
test cases and it is not clear what they're verifying. So instead of
testing something with an unclear rationale this commit just removes the
tests. We can always add them back if the purpose becomes clear in the
future.