Bump PyO3 version and support Stable Rust (#91)
* Bump PyO3 version and support Stable Rust

This commit updates retworkx to use the latest version of PyO3 which
enables support for compiling with stable rust (rustc >= 1.39). While
updating the PyO3 version supported this takes the opportunity to update
all the documentation and CI builds to rely on stable rust instead of
nightly. The nightly step was always an extra manual step that detracted
from the contribution process and prevented use in environments where
rustup wasn't allowed to be used.

* More README cleanup

* Bump PyO3 version to latest

* Tweak readme some more

* Fix readme whitespace

* Update README based on review comments

* Switch s390x wheel script to use latest stable too

s390x wheel jobs were recently added in #97 this commit updates the
install script for that job (rustup can't be used because of compat
issues with openssl in clefos7, which is used in the manylinux2014
docker image, and rustup) to pull the latest stable version of the
rust offline installer instead of nightly.

* Bump cibuildwheel version to the latest

* Tweak CI config to work with stable

* Fix typo in job name

* Explicitly use stable for cargo fmt job

* Try using travis to set rust version

* Explicitly use stable rust for cargo

* DNM: Test wheel builds

* Fix s390x job tarball filename

* Adjust wheel jobs again

* Revert "DNM: Test wheel builds"

This reverts commit 708e77145f44d7316aff05a4324a7642bc5bdd92.