Pivot to m2r2
m2r is broken on sphinx3 and the maintainer isn't responsive. Someone
has forked it and created a new package for m2r, called m2r2. The conda
rtd environment won't use sphinx2, it unconditionally upgrades as part
of it's standard steps before a build. So to workaround m2r this commit
switches to m2r2 instead.