Attempt to fix readthedocs build (#92)
* Attempt to fix readthedocs build

Since #90 the docs are now autogenerated from the installed retworkx
package. However, the readthedocs builds are not working because they
do not have rustc installed in their build env so it is unable to build
retworkx. This commit attempts to configure readthedocs to install rust
via conda and then use that to build retworkx.

* Pin sphinx to version that works with m2r

* Downgrade pip again needs to be <3 not <3.1

* Pivot to m2r2

m2r is broken on sphinx3 and the maintainer isn't responsive. Someone
has forked it and created a new package for m2r, called m2r2. The conda
rtd environment won't use sphinx2, it unconditionally upgrades as part
of it's standard steps before a build. So to workaround m2r this commit
switches to m2r2 instead.