Bumpy rust-numpy dependency version (#100)
The rust-numpy project pushed a new release right after 0.10.0 to fix
some soundness and safety bugs introduced in 0.10.0. These don't really
effect retworkx because we don't take numpy input from anything (yet)
and just use the bindings to wrap ndarrays as numpy.ndarray objects for
a return type. But, it's better to keep up-to-date as our dependencies
are updated. This commit bumps the rust-numpy dependency version in
Cargo.toml to the latest version.