Small docs cleanups
This commit fixes some small oversight in the docs. Primarily it adds
![text_signature] macros to the functions and methods that were missing
them. This is important mostly so that python's signature inspection is
aware of the arguments to the function. Sphinx doesn't always reliably
detect the signature from it. The other thing is a couple of missing
docstrings are added.