Configure dependabot to ignore fixedbitset for now
The fixedbitset dependency has to be kept in sync with the version
petgraph is using since we need to interoperate with pegraph using
fixedbitset objects a version mismatch will cause a compilation failure.
Right now dependabot is trying to update the fixedbitset verision
independently which will not pass CI. This commit sets fixedbitset as an
ignored package in the dependabot configuration.