Add to_dot method to PyGraph and PyDiGraph (#103)
* Add to_dot method to PyGraph and PyDiGraph

This commit adds a new method to PyGraph and PyDiGraph, to_dot(), which
is used to generate a graphviz dot file for a graph object. The
implementation of this is influenced by petgraph's built-in dot module
[1].

[1] https://github.com/petgraph/petgraph/blob/0.5.1/src/dot.rs

* Expand test coverage

* Add new doc requirements to rtd conda env

* Close before remove for windows

* Apply suggestions from code review

Co-authored-by: Lauren Capelluto <laurencapelluto@gmail.com>

* Make node_attrs and edge_attrs default to None

This commit changes the node_attrs and edge_attrs callables to be opt-in
and default to None. This makes the user burden lower because if you
don't require extra formatting this simplifies the call signature.

* Fix return types from to_dot methods

The to_dot methods were documented as returning a string if filename was
not specified, but we were returing a bytes object not a string. This
commit corrects the return type to convert the bytes array to a string
primitive and returning that instead of the raw PyBytes object.

Co-authored-by: Lauren Capelluto <laurencapelluto@gmail.com>