Use Kosaraju's algorithm instead of tarjan
The petgraph tarjan_scc() function is causing segfaults in some
scenarios. I'm not really sure if this is because of an issue with
petgraph or the combination of petgraph and PyO3 since both have
unstable blocks which could potentially lead to a segfault. Either way
switching the the implementation to use petgraph's kosaraju_scc function
solves this problem.