Add option to set multigraph on PyGraph constructor
This commit adds a new kwarg, multigraph, to the PyGraph constructor.
When this is set to False the created PyGraph object will not be a
multigraph and only one edge will exist between nodes. If a parallel
edge is attempted to be added it will be ignored since the edge already
exists.