Implement mapping protocol dunder methods for graph classes
This commit adds implementations for the __getitem__, __setitem__, and
__delitem__ mapping protocol methods to the PyGraph and PyDiGraph
classes. This enables accessing nodes from the graph by using standard
dict/mapping access patterns (ie graph[2]) with the node index. This
makes it easier to work with methods and functions that return node
indexes without having to call helper methods.