Add job testing the minimum supported rust version
We document that the minimum rust version we can compile with is 1.39.0.
However, we only ever test with the latest stable release (and nightly in
the case of the coverage job) so it's not clear when/if we introduce a
dependency on a newer version of the rust compiler to build retworkx.
This commit adds a new minimum supported rust version job that builds
retworkx with the MSRV 1.39.0 and runs tests on it that way we can catch
when something changes.