Remove leftover dist == 0 continue
There was a leftover check from an earlier iteration of the function
where the minimum dist was 0 because that was used to represent no edge.
But to function correctly this had to be change to use np.inf instead
making the check if 0 and skipping that location problematic. Having
this check doesn't break any of the test cases, but could interfere
with other graphs/weights especially if 0 is valid edge weight.