Add strongly connected components function (#111)
* Add strongly connected components function

This commit adds a function for computing the strongly connected
components from a PyDiGraph (and PyDAG) object. It returns a list of
lists of node ids of the strongly connected components.

* Add missing space to text_signature

* Use Kosaraju's algorithm instead of tarjan

The petgraph tarjan_scc() function is causing segfaults in some
scenarios. I'm not really sure if this is because of an issue with
petgraph or the combination of petgraph and PyO3 since both have
unstable blocks which could potentially lead to a segfault. Either way
switching the the implementation to use petgraph's kosaraju_scc function
solves this problem.

* Adjust assertion based on review comments