Use HashSet from hashbrown instead of std::collections
This commit switches the HashSet usage in retworkx to use hashbrown's
version of HashSet instead of the version in std::collections. This
provides a performance boost primarily from the use of the AHash default
hash algorithm. It also opens up potential future performance
improvements with hashbrown's rayon support to use parallel iterators.

At the same time since we're already listing rayon as a dependency
in #112 this adds the rayon feature to hashbrown in the cargo.toml
to expose the ParallelIterator traits for hashbrown's HashMap and
HashSet types. While this isn't being used yet it's potentially
useful in the future for #86.