Quote label strings in to_dot() output (#128)
This commit adds quotes around label attributes in the output from the
to_dot() methods. Previously the string was just passed raw to the
output, but this would cause issues in graphviz when trying to parse the
dot file if there were spaces or other restricted characters in the
string (eg having a label like '1 -> 2'. To fix this this commit puts a
quotes around a label string if its present to handle this case.