Add license to python package metadata (#132)
The license was not set for the python package metadata directly. We
listed the trove classifier for the project being apache licensed but
without specifying the license field certain parsers don't know what the
library is licensed under. This commit fixes the oversight and sets the
license field to apache 2.0.