Remove PyGraph implementation
The PyGraph implementation is a bit more involved since there is no
directionality to the edges. The implementation initially added didn't
take this into account and was basically just a copy of the PyDiGraph
implementation. But this was invalid and instead of adding a single edge
between each neighbor it would add 2 edges one with the weight for the
edge between the removed node and the neighbor. Since there is no
pressing use case for this on an undirected graph this commit just
removes the implementation, if a need arises in the future we can
revisit writing the implementation.