Remove common private method for building adjacency matrix
In earlier commits the adjacency matrix was built in private internal
functions that was originally just the code from the existing adjacency
matrix functions. However, as the floyd warshall numpy methods were
being developed a different form of adjacency matrix was needed where
the default value in the matrix was infinity and not 0. This made the
private function basically just 2 separate implementations in a single
function which lead to unecessary complexity. This commit breaks that
back out so that the infinity based approach is used for floyd warshall
and the 0 based approach is used for adjacency matrix.