Add remove_node_retain_edges method to PyDiGraph (#134)
* Add remove_node_retain_edges method to graph classes

This commit adds a new method remove_node_retain_edges() which is used
to remove a node but add edges from all predecessors to all descendants.

* Update src/digraph.rs

Co-authored-by: Kevin Krsulich <kevin@krsulich.net>

* Add options to choose which edge data to use and conditionally retain edges

* Remove PyGraph implementation

The PyGraph implementation is a bit more involved since there is no
directionality to the edges. The implementation initially added didn't
take this into account and was basically just a copy of the PyDiGraph
implementation. But this was invalid and instead of adding a single edge
between each neighbor it would add 2 edges one with the weight for the
edge between the removed node and the neighbor. Since there is no
pressing use case for this on an undirected graph this commit just
removes the implementation, if a need arises in the future we can
revisit writing the implementation.

Co-authored-by: Kevin Krsulich <kevin@krsulich.net>