Update exception usage in new PyO3 version
As documented in the changelog, release notes, and migration guide PyO3
0.12.0 reworked how exceptions are created and used. This was a breaking
change for how retworkx was using exceptions. This commit updates the
usage to be compatible with 0.12.0's new exception API. However, this is
only the first step on updating the PyO3 version and the compilation
still fails, this is currently blocked waiting for a rust-numpy release
that updates its usage to PyO3 0.12.0. Once that is released we can
continue this.