Handle invalid index first_layer arg for layers() func
This commit handles an error case in the retworkx.layers() algorithm
function when an invalid node index is part of the list passed in for
the 'first_layer' arg. Previously, this would result in a PanicException
being raised by PyO3 and include a backtrace from rust, this is because
the layers function was unconditionally pulling the node index from the
inner graph without checking whether it was a valid index or not. This
commit fixes this behavior and checks the data in the graph for that
index. If there is no data we raise an appropriate exception with an
exception message including the bad node index.