Add graph_floyd_warshall_numpy function for graph classes (#83)
* Add graph_floyd_warshall_numpy function for PyGraph

This commit adds a new function floyd_warshall_numpy() which is used to
find all-pairs of shortest path lengths. It returns the output as a
numpy matrix with shortest path distances between nodes. If there is no
path between nodes in the graph the value will be 0.

* Fix test failures and expand testing

* Add docs and fix lint

* Fix python lint and msrv failure

* Remove leftover dist == 0 continue

There was a leftover check from an earlier iteration of the function
where the minimum dist was 0 because that was used to represent no edge.
But to function correctly this had to be change to use np.inf instead
making the check if 0 and skipping that location problematic. Having
this check doesn't break any of the test cases, but could interfere
with other graphs/weights especially if 0 is valid edge weight.

* Add digraph_floyd_warshall_numpy too

* Remove common private method for building adjacency matrix

In earlier commits the adjacency matrix was built in private internal
functions that was originally just the code from the existing adjacency
matrix functions. However, as the floyd warshall numpy methods were
being developed a different form of adjacency matrix was needed where
the default value in the matrix was infinity and not 0. This made the
private function basically just 2 separate implementations in a single
function which lead to unnecessary complexity. This commit breaks that
back out so that the infinity based approach is used for floyd warshall
and the 0 based approach is used for adjacency matrix.

* Add test with negative cycles

* Add tests for graphs without edges

* Add tests for graphs with removed nodes

* Deduplicate code further