Use rust edition 2018 (#145)
Prior to this commit retworkx wasn't specifying an edition and was
implicitly using rust 2015. There wasn't any particular reason for this,
mainly just an ommision when the project was first created. This commit
updates the project to use the newer edition 2018 so we can get some
nice syntax features if needed.