Use contiguous C arrays instead of fortran layout (#157)
This commit moves away from using a column-major fortran layout for the
ndarrays generated in library functions that return a numpy array. A
common use case for these arrays is to pass it from numpy (which is
how they're returned) to another C extension. While numpy can natively
handle all sorts of memory layouts for arrays, not everything accessing
the array in memory can. To make this easier and to avoid copying
in end user programs this removes the .f() used to tell ndarray to use
a fortran layout when allocating the array. There wasn't ever really a
reason for this besides sloppy copy and paste when we started using
ndarray with rust-numpy.