Make weight_fn optional adjacency matrix fw numpy
This commit makes the weight_fn argument for graph_adjacency_matrix,
digraph_adjacency_matrix, graph_floyd_warshall_numpy, and
digraph_floyd_warshall_numpy optional. A new kwarg is added
default_weight (which defaults to 1.0) which can be used instead of
passing a callable. If weight_fn is not set the value of default_weight
will be used for all edges. Previously, a function returning a fixed
value would have to be used to accomplish this. In practice there was
not much overhead to just using something like 'lambda _: 1' as the
weight fn, but it was a bit of a clumsy interface.