Improve graph class docs
This commit improves the documentation for the graph class. THe previous
class documentation didn't really include any details on the usage of
the classes or how they work. Instead it was just pointing people to
petgraph saying retworkx's classes are built around petgraph assuming
that was enough for users to figure out how to use the class. This
commit reworks the documentation to better explain how the classes are
used and provide examples on how to use it.