Add distance matrix functions
This commit adds 2 new functions, graph_distance_matrix() and
digraph_distance_matrix(), which will generate a distance matrix for a
graph object. It is for use cases when all edges are unweighted or have
the same weight and is significantly faster than other algorithms
because of this assumption. It also operates in parallel if the size of
the graph justifies it.