Fix unpickling on empty graphs (#165)
In __setstate__ for PyDiGraph and PyGraph there was an implicit
assumption that there was always a node in the graph. However if an
empty graph object was being unpickled it would cause a panic because
at one point we tried to find the maximum node index which returns None
if there is no maximum and we unconditionally unwrapped it instead of
handling a None. This commit fixes this issue by returning early if the
list of node indices is empty.

Fixes #164