Remove unused base test class module (#169)
When unittests were first added a base test class module base.py was
added. The intent was that when the tests were running in CI we'd be
running under stestr and want to have the base fixtures for test method
timeout, and capturing stdout, stderr, and pylogging when the tests were
running in parallel. However, running unittests in parallel isn't
necessary because it takes < 1 second to run all ~300 unittests and
running with stestr takes 2-5 seconds because of overhead launching
multiple processes (all compared to ~2-4 minutes to compile retworkx)
so this base test class went unused. This commit removes that base.py
test class module, if in the future we do start running tests in
parallel using stestr we can revert this and add it back.