Revert PyDiGraph compose to use PyObject return
This commit reverts the change for the PyDiGraph.compose() method to use
a PyObject return type of a PyDict again. Doing the HashMap return
resulted in a double iteration over the output map, the first to convert
NodeIndex to a usize and the second in the PyO3 wrapper to convert the
HashMap to a PyDict in the python ffi. Returning a PyObject enables to
do the NodeIndex to usize conversion and PyDict generation at the same
time with a single loop over the contents.