Rely on PyO3 type conversion for HashMap and HashSet (#171)
* Rely on PyO3 type conversion for HashMap and HashSet

Since PyO3 0.12.0 the trait implementations for converting from
hashbrown's HashMap and HashSet types. [1] This commit leverage
this so we do not have to internally convert these objects to and from
python types.

[1] https://github.com/PyO3/pyo3/pull/1114

* Revert PyDiGraph compose to use PyObject return

This commit reverts the change for the PyDiGraph.compose() method to use
a PyObject return type of a PyDict again. Doing the HashMap return
resulted in a double iteration over the output map, the first to convert
NodeIndex to a usize and the second in the PyO3 wrapper to convert the
HashMap to a PyDict in the python ffi. Returning a PyObject enables to
do the NodeIndex to usize conversion and PyDict generation at the same
time with a single loop over the contents.

* Fix PyGraph.compose() too

* Revert to PyObject for graph_greedy_color

* Standardize naming on weight_transform_callable