Switch rng.gen_range() to Uniform.sample() in Gnp functions
This commit updates the gnp random graph functions to internally use
Uniform.sample() instead of rng.gen_range(). The rand docs for
gen_range() recommend the use of Uniform if you're going to getting
random values in a range more than once, as the gen_range() function is
optimized for the single sample case. [1][2]

[1] https://docs.rs/rand/0.7.3/rand/trait.Rng.html#method.gen_range
[2] https://docs.rs/rand/0.7.3/rand/distributions/uniform/struct.Uniform.html