Cap setuptools-rust version for python 3.5 (#190)
The recent setuptools-rust 0.11.4 release dropped support for python
3.5, but didn't expose a minimum python version when doing so. This
means that on python3.5 pip will install the latest incompatible version
and fail to build retworkx (see PyO3/setuptools-rust#85). To workaround
this issue this commit sets a version cap for setuptools-rust when using
python 3.5. After the upcoming 0.6.0 release we will drop support for
python 3.5 so this can be removed then.