Add distance matrix functions (#166)
* Add distance matrix functions

This commit adds 2 new functions, graph_distance_matrix() and
digraph_distance_matrix(), which will generate a distance matrix for a
graph object. It is for use cases when all edges are unweighted or have
the same weight and is significantly faster than other algorithms
because of this assumption. It also operates in parallel if the size of
the graph justifies it.

* Run cargo fmt

* Add docs

* Adjust default parallel_threshold from 100 to 300

* Use float for array instead of usize

* Add as_undirected argument to digraph_distance_matrix

* Update src/lib.rs

Co-authored-by: Lauren Capelluto <laurencapelluto@gmail.com>

* Add comment about parallelization to digraph function