Fix algorithm for dfs_edges
The dfs_edges algorithm previously was using petgraph's built in DFS
iterator for simplicity. However this was designed to emit nodes in
depth first manner and was lacking the necessary state to build the
edge list. This was caught in review where one of the tests was
incorrectly adding an edge that didn't exist to the output. This was
because the output from the iterator didn't return the parent with the
node so the algoirthm previously incorrectly assumed the previous node
in returned was from the parent, even though it was at a different
level. This commit fixes this by abandoning the DFS iterator and instead
manually doing it's own depth first traversal of the graph. This
provides the flexibility to track the parents with each child node so
that when we find unvisited children we know which parent to associate
in the output edge returned. This then updates the tests to fix the
incorrect expected values so we're asserting the correct behavior of the
function.