Use IntoNeighbors instead of IntoNeighborsDirected
In reworking the dfs algorithm the trait used to specify the neighbors
method be defined on the input graph was incorrectly change from
IntoNeighbors to IntoNeighborsDirected. While this works because there
is overlap in the methods provided by the trait it isn't accurate because
only the neighbors() method is used. This was from a previous local
iteration where I was using neighbors_directed(), but ended up not
relying on that and switching back to neighbors. This commit corrects
the trait required to make it clear what is required for any callers.