Only implement sequence protocol
Using the sequence protocol we can still get an implicit iterator by
just casting it on the python side. This will still get us the lazy type
conversion but simplify the api and also make the behavior more
consistent.

At the same time to ensure we're handling negative indices correctly a
test method is added to verify that a negative index access to the
sequence works as expected.