Add custom iterator class for BFS successors return (#185)
* Add custom iterator class for BFS successors return

This commit changes the return type of the bfs_successors function to be
a custom class BFSSuccessors. This new return class implements both the
sequence protocol and iterator protocol. This means that aside from
explicit type checking it should be backwards compatible with the list
being previously returned. It can be used with either index based access
or iterated over.

This should be more efficient for large graphs because instead of doing
the copy and type conversion and iterating over the entire nested Vec of
results it instead does it per access (either via __getitem__ or
__next__). It does add a small amount of overhead for smaller graphs but
it is minimal since the function returns in microseconds in such cases
so a 10-20% overhead is not a big deal.

It's worth noting while this defers the type conversion, it does not
defer execution like most python iterators normally do. When
bfs_successors is called it will still always fully traverse the graph.
However, in practice the bottleneck for the bfs_successor function
wasn't actually the graph traversal, but instead the type conversion.

Related to #71

* Only implement sequence protocol

Using the sequence protocol we can still get an implicit iterator by
just casting it on the python side. This will still get us the lazy type
conversion but simplify the api and also make the behavior more
consistent.

At the same time to ensure we're handling negative indices correctly a
test method is added to verify that a negative index access to the
sequence works as expected.

* Update src/lib.rs

Co-authored-by: Kevin Krsulich <kevin@krsulich.net>