Add NodeIndices return type
This commit adds a new return type NodeIndices that is used as the
return type for functions that return Vec<usize> when its used as the
return type to return a list of node indices. This custom type implements
the Python sequence protocol and can be used as the list which was
previously returned except for where explicit type checking was used.
Just as in #185 the advantage here is that for large lists of node
indices the type conversion from a Vec<usize> to list(int) becomes a
large bottleneck. This avoids that conversion and only does a usize to
python int conversion on access.

Related to #71