Implement rich comparison for BFSSuccessors
This commit implements the rich comparison method of the python object
protocol for the BFSSuccessors class. This enables doing an equals and
not equals comparison between a BFSSuccessors object and any other
python sequence. If will compare all the elements of the 2 and return
appropriate result. Previously this was not implemented so you would
have to cast a BFSSuccessors object as a list first to compare it which
would iterate over the object. Implementing the method avoids the need
to do this.