Expand unit test coverage for PyDiGraph.merge_nodes()
In #131 the merge_nodes() method was added as part of the larger feature
being added. In that PR though there wasn't any testing of the error
conditions or when the node weights were different. This commit covers
those edge cases to fix the coverage gap and enforce the current
behavior moving forward.