Fix non-x86 release jobs
During the 0.6.0 release the travis jobs for building the release wheels
on aarch64, ppc64le, and s390x all failed because of some requirements
conflicts between the travis python environment and the requirements for
twine. The wheels were built successfully but the upload to PyPi failed
because of this requirements issue. This commit fixes that so future
releases should be able to upload the wheels. At the same time the rust
version used for the s390x jobs are updated is updated to the latest
stable release to keep the compiled binaries using the same version of
rustc.