Add EdgeList and WeightedEdgeList return types
This commit adds 2 new return types EdgeList and WeightedEdgeList
that are used as the return type for functions that return
Vec<(usize, usize)> and Vec<(usize, usize, PyObject)> respectively.
This custom type implements the Python sequence protocol and can be
used as the list which was previously returned except for where
explicit type checking was used. Just as in #185 and #198 the
advantage here is that for large edge lists the conversion from
Rust to a Python type becomes a large bottleneck. This avoids that
conversion and only does it per element on access.