Switch HashMap implementation in astar to use hashbrown (#205)
In the astar.rs module it keeps a couple of HashMaps to track the path
and scores as the algorithm is going through the graph. Those HashMaps
were using the std collections HashMap type. However, in the rest of
retworkx the hashbrown HashMap is used. This is because the hashbrown
version is faster by default (because it uses a non-DOS resistant hash
function which is faster) and also supports parallelism. This commit
updates the astar.rs module to be consistent with the rest of retworkx
and use hashbrown, which will improve performance of the function.