Update package, cap ndarray, and include cargo lock in sdist (#207)
This commit fixes a couple of packaging bugs with the current release of
retworkx. The first issues is that ndarray just released 1.14.0 which
raised the minimum supported rust version to 1.42.0 which is
incompatible with retworkx's minimum of 1.39.0. This caps the ndarray
package to only a compatible version with 0.13. The second issue was the
Cargo.lock was not included in the sdist. This means that when pepople
go to build from sdist via pypi (or anywhere else) the reproducible
build versions of all the dependencies are not used. This can cause
issues like the ndarray incompatibility which is blocking source builds
right now. The last thing is the cargo lock file is updated to pull in
the latest version of pyo3, 0.12.4 which fixes a potential security
issue with the earlier 0.12.x releases.