Handle multiple edges to a single successor node
There was a bug in the collect_runs() method when there were multiple
edges between a graph and its single successor node. Since
neighbors_directed() returns an iterator over the edges and not the
nodes it would have multiple entries in the succesors list despite only
being a single node. This commit addresses that edge case by building a
HashSet of the node indices in the succesors list and checking the
length of that instead of using the raw successors list. That will
contain no duplicate nodes and will identify if there is a path instead
of splitting on multiple edges incorrectly.