Fix documentation and signature issues
This commit fixes some issues with the documentation and text_signature
attribute for methods. The documentation of digraph_union needed a bit
of cleanup and more details to explain how it works. The other issue
fixed here is the text_signature attribute for methods were missing the
'self' parameter. While the self is implicit in the typical call,
without it in the text_signature attribute the sphinx builds assume the
first parameter in the signature is self, and removes it from the
documentation. This commit fixes this by adding self to all the method
signature attributes.