Document and test nodes can only exist in a single run
This commit documents and adds tests to assert that a node can only be
in a single run in the output from the collect_runs() method. In terra
the equivalent python function didn't have this additional constraint
so this commit attempts to make it clear so the differences between
the two implementations are understood by users.