Add collect_runs() function (#206)
* Add collect_runs() function

This commit adds a new algorithms function, collect_runs(), which is
used to find a list of runs in a PyDiGraph object that match a specified
condition. A run in this context is any linear path inside the graph.
The condition is specified using a python function that will be passed a
node's data payload/weight object and will return whether it matches the
condition or not. The intended use case for this function is with
qiskit-terra's PyDiGraph method collect_runs() which is implementing
this algorithm currently in Python on top of PyDiGraph (it will also be
used by a future method collect_1q_runs which is being added in
Qiskit/Qiskit-terra#5431).

* Handle multiple edges to a single successor node

There was a bug in the collect_runs() method when there were multiple
edges between a graph and its single successor node. Since
neighbors_directed() returns an iterator over the edges and not the
nodes it would have multiple entries in the succesors list despite only
being a single node. This commit addresses that edge case by building a
HashSet of the node indices in the succesors list and checking the
length of that instead of using the raw successors list. That will
contain no duplicate nodes and will identify if there is a path instead
of splitting on multiple edges incorrectly.

* Add initial tests

* Add more tests

* Use Vec.dedup() and graph access instead of Vec of tuples and HashSet

* Update src/lib.rs

* Apply suggestions from code review

Co-authored-by: Lauren Capelluto <laurencapelluto@gmail.com>

* Document and test nodes can only exist in a single run

This commit documents and adds tests to assert that a node can only be
in a single run in the output from the collect_runs() method. In terra
the equivalent python function didn't have this additional constraint
so this commit attempts to make it clear so the differences between
the two implementations are understood by users.

Co-authored-by: Lauren Capelluto <laurencapelluto@gmail.com>
Co-authored-by: Toshinari Itoko <itoko@jp.ibm.com>