Update documentation and add release notes
This commit updates the documentation theme to match the current theme
that the qiskit project is using. It also takes the opportunity to add
the release notes to the documentation. We previously were only
publishing the release documentation on the github release page which
has limitations. Moving forward we'll use reno to automate the
generation of release notes.